Make podcast openable in the browser.
We used to specify a MIME type of `audio/*` which prompted it to open in
a media player, but Firefox and Chrome did not appear in the picker. If
we don't specify a MIME type at all, just the URL, the browsers appear
in the picker (and actually VLC and Podcast Addict also pick up the
intent--maybe based on a .mp3 extension? dunno.)